Upgrade Kotlin to 1.3-RC, migrate coroutines. Add Architecture components.
Kotlin 1.3 brings with it a change to coroutines, pulling them out of "experimantal" and adding mandatory CoroutineScope requirements.
To make coroutines work throughout configuration changes (screen rotation), ViewModels should be used when fetching data from the internet. Thus, GameInfoFragment and SplitsFragment now each have a viewModel attached to them, taking care of network requests and proper UI handling when results arrive. The "kotlin-coroutines-retrofit" library is not really needed (and is not compatible with the new coroutines) and was removed.