Update Kotlin to 1.3.0, add flag to close timer broadcast intent
Extra boolean flag "fromOnResume" added to the closeTimer broadcast intent whether the timer is being closed from opening the app or from the notification. If it is being closed from opening the app, the "timerActiveDialog" must not be cancelable, otherwise the timer can still be displayed while using the app. Also, the dialog should minimize the app when hitting the cancel button. If the timer is being closed from the notification, no such restrictions need to be applied.